Serde evm sig with 0x prefix
