+++ tee-worker/omni-executor/executor-primitives/Cargo.toml
@@ -16,6 +16,7 @@ tracing = { workspace = true }
 
 executor-crypto = { workspace = true }
 heima-primitives = { workspace = true }
+heima-utils = { workspace = true }
 sp-core = { workspace = true, features = ["serde"] }
 
 [dev-dependencies]
