+++ tee-worker/omni-executor/executor-primitives/src/signature/mod.rs
@@ -1,8 +1,8 @@
 mod bitcoin;
 pub use bitcoin::BitcoinSignature;
 
-mod ethereum;
-pub use ethereum::EthereumSignature;
+mod evm;
+pub use evm::EthereumSignature;
 
 mod heima_multi_signature;
 pub use heima_multi_signature::{HeimaMultiSignature, SolanaSignature, SubstrateSignature};
